backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.5
celery==5.4.0
imagehash==4.3.1
Pillow==10.3.0
cryptography==42.0.8
python-dotenv==1.0.1
stripe==9.9.0
resend==2.0.0
psutil==5.9.8
aiofiles==23.2.1

# ── Month 2: Cloud OAuth + Thumbnails ─────────────────────────────────────
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.132.0
dropbox==12.0.2
boto3==1.34.131          # For R2/S3 thumbnail storage
python-magic==0.4.27     # MIME type detection from bytes

# ── Month 3: AI Classification + Billing ──────────────────────────────────
numpy==1.26.4             # Blur detection (Laplacian variance)
opencv-python-headless==4.9.0.80   # OpenCV for blur/screenshot detection
openai==1.35.3            # GPT-4o Vision for document classification (Pro)
msal==1.29.0              # Microsoft OAuth for OneDrive
APScheduler==3.10.4       # Scheduled cleanup jobs

# ── Month 4: CLIP Embeddings + Celery Workers + Email ─────────────────────
pgvector==0.3.2           # pgvector SQLAlchemy extension
celery==5.4.0             # Already listed — task queue (verified)
flower==2.0.1             # Celery monitoring UI
resend==2.0.0             # Already listed — email sending (verified)
jinja2==3.1.4             # HTML email templates

# ── Month 5: Export + Webhooks ────────────────────────────────────────────
openpyxl==3.1.4           # Excel export for file index
pandas==2.2.2             # DataFrame for reports/exports
PyJWT==2.8.0              # API key signing

# ── Month 6: PWA + Advanced ───────────────────────────────────────────────
qrcode==7.4.2             # QR codes for share links
python-slugify==8.0.4     # URL slug generation
